orders/src/events/listeners/__test__/ticket-updated-listener.test.ts
import { Ticket } from '../../../models/ticket';
import { natsWrapper } from '../../../nats-wrapper';
import {
  generateMessageData,
  generateTicketData,
  generateTicketUpdatedEvent,
} from '../../../test/test-data-generators';
import { TicketUpdatedListener } from '../ticket-updated-listener';

const setup = async () => {
  const listener = new TicketUpdatedListener(natsWrapper.client);
  const ticket = Ticket.build(generateTicketData());
  await ticket.save();
  const data = generateTicketUpdatedEvent(ticket);
  const msg = generateMessageData();

  return { msg, data, ticket, listener };
};

it('finds, updates, and saves a ticket', async () => {
  const { msg, data, ticket, listener } = await setup();
  await listener.onMessage(data, msg);

  const updatedTicket = await Ticket.findById(ticket.id);
  expect(updatedTicket!.title).toEqual(data.title);
  expect(updatedTicket!.price).toEqual(data.price);
  expect(updatedTicket!.version).toEqual(data.version);
});

it('acks the message', async () => {
  const { msg, data, listener } = await setup();

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});

it('does not call ack if the event hasa skipped version number', async () => {
  const { msg, data, listener, ticket } = await setup();
  data.version = 10;

  try {
    await listener.onMessage(data, msg);
  } catch (err) {}
  expect(msg.ack).not.toHaveBeenCalled();
});
